Tuesday PM

+++ sql/init.sql
@@ -39,7 +39,7 @@ CREATE INDEX idx_pet_customer ON pets(customer_id);
 CREATE TABLE services (
     service_id INT AUTO_INCREMENT PRIMARY KEY,
     service_name VARCHAR(100) NOT NULL,
-    service_price DECIMAL(10,2) NOT NULL,
+    service_price INT(11) NOT NULL,
     pet_type VARCHAR(100) NOT NULL,
     pet_size VARCHAR(100) NOT NULL
 );
